# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FreightUtils API",
    "description": "Free freight calculation and reference data APIs. No authentication required. JSON responses. CORS enabled for all origins.",
    "version": "1.0.0",
    "contact": {
      "name": "FreightUtils",
      "url": "https://www.freightutils.com",
      "email": "contact@freightutils.com"
    },
    "license": {
      "name": "Free to use",
      "url": "https://www.freightutils.com/api-docs"
    }
  },
  "servers": [
    {
      "url": "https://www.freightutils.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/containers": {
      "get": {
        "operationId": "getContainers",
        "summary": "Container Capacity Reference",
        "description": "Shipping container specifications for all 10 standard ISO container types. Returns dimensions, weights, door openings, and pallet capacity. Optionally calculate how many items of given dimensions fit in a specific container.",
        "tags": [
          "Reference Data"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Container type slug. Omit to list all containers.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "20ft-standard",
                "40ft-standard",
                "20ft-high-cube",
                "40ft-high-cube",
                "20ft-open-top",
                "40ft-open-top",
                "20ft-flat-rack",
                "40ft-flat-rack",
                "20ft-reefer",
                "40ft-reefer"
              ]
            },
            "example": "40ft-high-cube"
          },
          {
            "name": "l",
            "in": "query",
            "description": "Item length in centimetres. Requires type, w, and h for loading calculation.",
            "required": false,
            "schema": {
              "type": "number"
            },
            "example": 60
          },
          {
            "name": "w",
            "in": "query",
            "description": "Item width in centimetres.",
            "required": false,
            "schema": {
              "type": "number"
            },
            "example": 40
          },
          {
            "name": "h",
            "in": "query",
            "description": "Item height in centimetres.",
            "required": false,
            "schema": {
              "type": "number"
            },
            "example": 40
          },
          {
            "name": "wt",
            "in": "query",
            "description": "Item weight in kilograms.",
            "required": false,
            "schema": {
              "type": "number"
            },
            "example": 15
          },
          {
            "name": "qty",
            "in": "query",
            "description": "Number of items.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "example": 100
          }
        ],
        "responses": {
          "200": {
            "description": "Container specifications or loading calculation result.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ContainerListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ContainerSpec"
                    },
                    {
                      "$ref": "#/components/schemas/ContainerLoadingResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Container type not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consignment": {
      "post": {
        "summary": "Multi-item consignment calculator",
        "description": "Calculate total CBM, weight, LDM, and chargeable weight for mixed consignments. Supports road, air, and sea modes.",
        "operationId": "calculateConsignment",
        "tags": [
          "Consignment"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "items"
                ],
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "road",
                      "air",
                      "sea"
                    ],
                    "default": "road"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "length",
                        "width",
                        "height"
                      ],
                      "properties": {
                        "description": {
                          "type": "string"
                        },
                        "length": {
                          "type": "number"
                        },
                        "width": {
                          "type": "number"
                        },
                        "height": {
                          "type": "number"
                        },
                        "quantity": {
                          "type": "integer",
                          "default": 1
                        },
                        "grossWeight": {
                          "type": "number"
                        },
                        "stackable": {
                          "type": "boolean"
                        },
                        "pallet_type": {
                          "type": "string",
                          "enum": [
                            "none",
                            "euro",
                            "uk",
                            "us"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consignment calculation result"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/duty": {
      "post": {
        "summary": "UK import duty and VAT estimator",
        "description": "Estimate UK import duty and VAT using live GOV.UK Trade Tariff data.",
        "operationId": "estimateDuty",
        "tags": [
          "Customs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "commodity_code",
                  "origin_country",
                  "customs_value"
                ],
                "properties": {
                  "commodity_code": {
                    "type": "string",
                    "example": "6204430000"
                  },
                  "origin_country": {
                    "type": "string",
                    "example": "CN"
                  },
                  "customs_value": {
                    "type": "number",
                    "example": 10000
                  },
                  "freight_cost": {
                    "type": "number",
                    "example": 500
                  },
                  "insurance_cost": {
                    "type": "number",
                    "example": 50
                  },
                  "incoterm": {
                    "type": "string",
                    "example": "FOB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duty and VAT estimate"
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message.",
            "example": "Missing required parameter: length"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ContainerSpec": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Container type name.",
            "example": "40ft High Cube"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier.",
            "example": "40ft-high-cube"
          },
          "internal_length_cm": {
            "type": "number",
            "description": "Internal length in cm.",
            "example": 1203
          },
          "internal_width_cm": {
            "type": "number",
            "description": "Internal width in cm.",
            "example": 235
          },
          "internal_height_cm": {
            "type": "number",
            "description": "Internal height in cm.",
            "example": 269
          },
          "external_length_cm": {
            "type": "number",
            "description": "External length in cm.",
            "example": 1219
          },
          "external_width_cm": {
            "type": "number",
            "description": "External width in cm.",
            "example": 244
          },
          "external_height_cm": {
            "type": "number",
            "description": "External height in cm.",
            "example": 290
          },
          "door_width_cm": {
            "type": "number",
            "nullable": true,
            "description": "Door opening width in cm.",
            "example": 234
          },
          "door_height_cm": {
            "type": "number",
            "nullable": true,
            "description": "Door opening height in cm.",
            "example": 263
          },
          "capacity_cbm": {
            "type": "number",
            "description": "Internal volume in CBM.",
            "example": 76.2
          },
          "tare_weightKg": {
            "type": "number",
            "description": "Empty container weight in kg.",
            "example": 3900
          },
          "max_gross_kg": {
            "type": "number",
            "description": "Maximum gross weight in kg.",
            "example": 30480
          },
          "max_payload_kg": {
            "type": "number",
            "description": "Maximum cargo weight in kg.",
            "example": 26580
          },
          "euro_pallets": {
            "type": "string",
            "description": "Number of EUR pallets (1200x800mm) that fit.",
            "example": "25"
          },
          "gma_pallets": {
            "type": "string",
            "description": "Number of GMA pallets (48x40in) that fit.",
            "example": "20"
          },
          "description": {
            "type": "string",
            "description": "Container description."
          },
          "notes": {
            "type": "string",
            "description": "Usage notes."
          }
        },
        "required": [
          "name",
          "slug",
          "internal_length_cm",
          "internal_width_cm",
          "internal_height_cm",
          "capacity_cbm",
          "tare_weightKg",
          "max_gross_kg",
          "max_payload_kg",
          "euro_pallets",
          "gma_pallets"
        ]
      },
      "ContainerListResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of containers.",
            "example": 10
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContainerSpec"
            }
          }
        },
        "required": [
          "count",
          "results"
        ]
      },
      "ContainerLoadingResponse": {
        "type": "object",
        "properties": {
          "container": {
            "$ref": "#/components/schemas/ContainerSpec"
          },
          "loading": {
            "type": "object",
            "properties": {
              "fits_lengthwise": {
                "type": "integer",
                "description": "Items fitting along the length."
              },
              "fits_widthwise": {
                "type": "integer",
                "description": "Items fitting along the width."
              },
              "fits_height": {
                "type": "integer",
                "description": "Items fitting in height."
              },
              "max_items": {
                "type": "integer",
                "description": "Maximum items that physically fit."
              },
              "requested_qty": {
                "type": "integer",
                "description": "Quantity requested."
              },
              "fits": {
                "type": "boolean",
                "description": "Whether all requested items fit."
              },
              "total_weight_kg": {
                "type": "number",
                "description": "Total weight of loaded items."
              },
              "within_payload": {
                "type": "boolean",
                "description": "Whether total weight is within payload limit."
              },
              "utilisation_percent": {
                "type": "number",
                "description": "Volume utilisation percentage."
              }
            }
          }
        },
        "required": [
          "container",
          "loading"
        ]
      },
...
    }
  }
}